Move auto-send worker and upload results to upload package
